Extract shell socket server
